+++ packages/web3-bzz/src/index.js
@@ -28,6 +28,10 @@ var swarm = require("swarm-js");
 
 var Bzz = function Bzz(provider) {
 
+
+    console.warn('web3-bzz package will be deprecated in version 1.3.4 and will no longer be supported.')
+
+
     this.givenProvider = Bzz.givenProvider;
 
     if (provider && provider._requestManager) {