+++ CHANGELOG.md
@@ -327,19 +327,21 @@ Released with 1.0.0-beta.37 code base.
 
 ### Changed
 
-- Remove `notImplemented` flag from ETH2 Beacon Chain package methods schema (#3861)
 - Fixed mutation of inputs to encoding and decoding functions (#3748)
+- Fix default value for `fromBlock` option for `logs` subscriptions (defaults to `latest`) (#3883)
+- ethjs-signer test (#3876)
 - Rename `web3-eth2-base` to `web3-eth2-core` and `web3-eth2-beacon` to `web3-eth2-beaconchain` (#3833)
 - Bump `ts-node` from version `^8.10.2` to `^9.0.0` (#3856)
 - Ran `npm audit fix` which fixed 4 vulnerabilities (#3856)
 - Correct `web3-eth2-beaconchain` type declarations (#3859) and (#3865)
 - Move interfaces `IBaseAPISchema` and `IBaseAPIMethodSchema` to `index.d.ts` for `web3-eth2-core` (#3878)
-- Removes `IETH2BeaconChain` interface in favor of exporting a class type: `ETH2BeaconChain` (#3878)
-- Remove `index.d.ts` files in favor of `types.ts` for `web3-eth2-core` and `web3-eth2-beaconchain` (#3878)
 - Update dependencies for `web3-eth2-core` (#3878)
 
 ### Removed
 
+- Remove `notImplemented` flag from ETH2 Beacon Chain package methods schema (#3861)
+- Removes `IETH2BeaconChain` interface in favor of exporting a class type: `ETH2BeaconChain` (#3878)
+- Remove `index.d.ts` files in favor of `types.ts` for `web3-eth2-core` and `web3-eth2-beaconchain` (#3878)
 - `schema.ts` from `web3-eth2-core` (#3878)
 - `dtslint` npm command from `web3-eth2-core` and `web3-eth2-beaconchain` as `index.d.ts` files were removed (#3878)