+++ CHANGELOG.md
@@ -321,8 +321,6 @@ Released with 1.0.0-beta.37 code base.
 
 - Update `dist` to latest release (`1.3.2`) (#3875)
 
-## [Unreleased]
-
 ## [1.3.4]
 
 ### Changed
@@ -349,3 +347,7 @@ Released with 1.0.0-beta.37 code base.
 
 - Add `ETH2Core` class export to `index.d.ts` for `web3-eth2-core` (#3878)
 - Deprecation of bzz warning (#3872)
+
+## [Unreleased]
+
+## [1.3.5]